a bunch of annoying gcc 2.9x compatibility crap

+++ util/HexDumpBuffer.cc
@@ -91,7 +91,7 @@ HexDumpBuffer::hexify()
         append("  ");
     }
 
-    appendf(" |  %.*s\n", len % 16, strbuf);
+    appendf(" |  %.*s\n", (int)len % 16, strbuf);
 }
 
 } // namespace oasys
